+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 2026-06-06
 
+- Added Ansible primer (`Ansible/notes/0000-primer-ansible.md`)
+- Added GitHub primer (`GitHub/notes/0000-primer-github.md`)
+- Added Terraform install + init script (`Terraform/scripts/install-and-init.sh`)
 - Added Docker install + first container script (`Docker/scripts/install-and-run-first-container.sh`)
 - Added Docker CLI exploration notes (`Docker/notes/2026-06-06-exploring-docker-cli.md`)
 - Added Kubernetes primer (`Kubernetes/notes/0000-primer-kubernetes.md`)